Time: 405 ms (7.86%), Space: 56.9 MB (5.24%) - LeetHub

+++ 2288-apply-discount-to-prices/2288-apply-discount-to-prices.java
@@ -18,8 +18,8 @@ class Solution {
     }
 
     private String discountString(String word, int discount) {
-        int price = Integer.parseInt(word.substring(1));
-        double discountedPrice = (double)price*(100-discount)/100;
+        double price = Double.parseDouble(word.substring(1));
+        double discountedPrice = price*(100-discount)/100;
         System.out.println(price + " " + discountedPrice);
         return "$"+String.format("%.2f", discountedPrice);
     }
